Use floats for calculation.

+++ api.go
@@ -2,7 +2,7 @@ package gologo
 
 func CreateCircle(x, y int32) {
     objects[next_object_id] = &Circle{
-        Centre: Vector{ x: x, y: y },
+        Centre: Vector{ x: float64(x), y: float64(y) },
         Radius: 20,
         Renderer: renderers[RENDER_OBJ],
     }
